+++ workspaces/topology/package.json
@@ -9,7 +9,7 @@
     "tsc": "tsc",
     "tsc:full": "tsc --skipLibCheck true --incremental false",
     "start": "backstage-cli repo start",
-    "start:alpha": "yarn workspaces foreach -A --include @backstage-community/plugin-topology --include @internal/plugin-topology-dev-backend --parallel -v -i run start:alpha",
+    "start:legacy": "yarn workspaces foreach -A --include @backstage-community/plugin-topology --include @internal/plugin-topology-dev-backend --parallel -v -i run start:legacy",
     "build:all": "backstage-cli repo build --all",
     "build:api-reports": "yarn build:api-reports:only",
     "build:api-reports:only": "backstage-repo-tools api-reports -o ae-wrong-input-file-type,ae-undocumented --validate-release-tags",
@@ -17,9 +17,9 @@
     "clean": "backstage-cli repo clean",
     "test": "backstage-cli repo test",
     "test:all": "backstage-cli repo test --coverage",
+    "test:nfs": "APP_MODE=nfs playwright test",
     "test:legacy": "APP_MODE=legacy playwright test",
-    "test:alpha": "APP_MODE=alpha playwright test",
-    "test:e2e:ci": "yarn test:legacy && yarn test:alpha",
+    "test:e2e:ci": "yarn test:nfs && yarn test:legacy",
     "playwright": "sh -c 'if [ \"$1\" = test ] && [ $# -eq 1 ]; then yarn test:e2e:ci; else exec playwright \"$@\"; fi' _",
     "fix": "backstage-cli repo fix",
     "lint": "backstage-cli repo lint --since origin/main",
