+++ workspaces/topology/plugins/topology-dev-backend/package.json
@@ -19,7 +19,7 @@
   },
   "scripts": {
     "start": "backstage-cli package start",
-    "start:alpha": "backstage-cli package start",
+    "start:legacy": "backstage-cli package start",
     "build": "backstage-cli package build",
     "lint": "backstage-cli package lint",
     "clean": "backstage-cli package clean"
