+++ workspaces/topology/plugins/topology/app-config.backstage-community.yaml
@@ -4,6 +4,7 @@ dynamicPlugins:
       mountPoints:
         - mountPoint: entity.page.topology/cards
           importName: TopologyPage
+          module: Legacy
           config:
             layout:
               gridColumn: '1 / -1'
