--- workspaces/topology/plugins/topology/dev/alpha/index.mock.tsx
@@ -1,142 +1,0 @@
-/*
- * Copyright 2026 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * New Frontend System dev mode for the Topology plugin (mock data).
- */
-
-import '@backstage/cli/asset-types';
-// eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
-import '@backstage/ui/css/styles.css';
-
-import ReactDOM from 'react-dom/client';
-
-import { createApp } from '@backstage/frontend-defaults';
-import {
-  ApiBlueprint,
-  createFrontendModule,
-  pluginHeaderActionsApiRef,
-} from '@backstage/frontend-plugin-api';
-import { catalogApiRef } from '@backstage/plugin-catalog-react';
-import {
-  kubernetesApiRef,
-  kubernetesAuthProvidersApiRef,
-} from '@backstage/plugin-kubernetes-react';
-import { AuthorizeResult } from '@backstage/plugin-permission-common';
-import { permissionApiRef } from '@backstage/plugin-permission-react';
-
-import topologyPlugin from '../../src/alpha';
-import topologyTranslationsModule from '../../src/translations/exports';
-import { devSidebarContent } from './shared';
-
-import {
-  mockCatalogApi,
-  mockKubernetesClient,
-  mockKubernetesAuthProviderApi,
-} from '../mocks';
-
-const catalogDevModule = createFrontendModule({
-  pluginId: 'catalog',
-  extensions: [
-    ApiBlueprint.make({
-      params: defineParams =>
-        defineParams({
-          api: catalogApiRef,
-          deps: {},
-          factory: () => mockCatalogApi,
-        }),
-    }),
-  ],
-});
-
-const kubernetesMockModule = createFrontendModule({
-  pluginId: 'kubernetes',
-  extensions: [
-    ApiBlueprint.make({
-      params: defineParams =>
-        defineParams({
-          api: kubernetesApiRef,
-          deps: {},
-          factory: () => mockKubernetesClient,
-        }),
-    }),
-    ApiBlueprint.make({
-      name: 'auth-providers',
-      params: defineParams =>
-        defineParams({
-          api: kubernetesAuthProvidersApiRef,
-          deps: {},
-          factory: () => mockKubernetesAuthProviderApi,
-        }),
-    }),
-  ],
-});
-
-const appDevModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [
-    ApiBlueprint.make({
-      name: 'permission',
-      params: defineParams =>
-        defineParams({
-          api: permissionApiRef,
-          deps: {},
-          factory: () => ({
-            authorize: async () => ({
-              result: window.location.pathname.includes('permission-denied')
-                ? AuthorizeResult.DENY
-                : AuthorizeResult.ALLOW,
-            }),
-          }),
-        }),
-    }),
-    ApiBlueprint.make({
-      name: 'plugin-header-actions',
-      params: defineParams =>
-        defineParams({
-          api: pluginHeaderActionsApiRef,
-          deps: {},
-          factory: () => ({
-            getPluginHeaderActions: () => [],
-          }),
-        }),
-    }),
-  ],
-});
-
-const devNavModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [devSidebarContent],
-});
-
-const app = createApp({
-  features: [
-    devNavModule,
-    topologyPlugin,
-    topologyTranslationsModule,
-    catalogDevModule,
-    kubernetesMockModule,
-    appDevModule,
-  ],
-});
-
-if (window.location.pathname === '/') {
-  window.location.replace('/catalog');
-}
-
-const root = app.createRoot();
-
-ReactDOM.createRoot(document.getElementById('root')!).render(root);
