+++ workspaces/topology/plugins/topology/dev/legacy.mock.tsx
@@ -1,0 +1,102 @@
+/*
+ * Copyright 2026 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Legacy frontend system dev mode for the Topology plugin (mock data).
+ */
+
+import { createDevApp } from '@backstage/dev-utils';
+import { Page, Header, TabbedLayout } from '@backstage/core-components';
+import { EntityProvider } from '@backstage/plugin-catalog-react';
+import { mockApis, TestApiProvider } from '@backstage/test-utils';
+import { AuthorizeResult } from '@backstage/plugin-permission-common';
+import { permissionApiRef } from '@backstage/plugin-permission-react';
+import {
+  kubernetesApiRef,
+  kubernetesAuthProvidersApiRef,
+} from '@backstage/plugin-kubernetes-react';
+import { TopologyPage, topologyPlugin } from '../src/legacy';
+import { topologyTranslations } from '../src/translations';
+import {
+  mockEntity,
+  mockKubernetesAuthProviderApi,
+  mockKubernetesClient,
+  permissionDeniedMockEntity,
+} from './mocks';
+
+createDevApp()
+  .registerPlugin(topologyPlugin)
+  .addTranslationResource(topologyTranslations)
+  .setAvailableLanguages(['en', 'de', 'fr', 'it', 'es', 'ja'])
+  .setDefaultLanguage('en')
+  .addPage({
+    element: (
+      <TestApiProvider
+        apis={[
+          [kubernetesApiRef, mockKubernetesClient],
+          [kubernetesAuthProvidersApiRef, mockKubernetesAuthProviderApi],
+          [permissionApiRef, mockApis.permission()],
+        ]}
+      >
+        <EntityProvider entity={mockEntity}>
+          <Page themeId="service">
+            <Header
+              type="component — service"
+              title={mockEntity.metadata.name}
+            />
+            <TabbedLayout>
+              <TabbedLayout.Route path="/" title="Topology">
+                <TopologyPage />
+              </TabbedLayout.Route>
+            </TabbedLayout>
+          </Page>
+        </EntityProvider>
+      </TestApiProvider>
+    ),
+    title: 'Topology',
+    path: '/topology',
+  })
+  .addPage({
+    element: (
+      <TestApiProvider
+        apis={[
+          [kubernetesApiRef, mockKubernetesClient],
+          [kubernetesAuthProvidersApiRef, mockKubernetesAuthProviderApi],
+          [
+            permissionApiRef,
+            mockApis.permission({ authorize: AuthorizeResult.DENY }),
+          ],
+        ]}
+      >
+        <EntityProvider entity={permissionDeniedMockEntity}>
+          <Page themeId="service">
+            <Header
+              type="component — service"
+              title={permissionDeniedMockEntity.metadata.name}
+            />
+            <TabbedLayout>
+              <TabbedLayout.Route path="/" title="Topology">
+                <TopologyPage />
+              </TabbedLayout.Route>
+            </TabbedLayout>
+          </Page>
+        </EntityProvider>
+      </TestApiProvider>
+    ),
+    title: 'Missing permissions',
+    path: '/missing-permissions',
+  })
+  .render();
