+++ workspaces/topology/plugins/topology/knip-report.md
@@ -4,9 +4,9 @@
 
 | Name                            | Location          | Severity |
 | :------------------------------ | :---------------- | :------- |
-| @backstage/plugin-user-settings | package.json:72:6 | error    |
-| @backstage/filter-predicates    | package.json:62:6 | error    |
-| @backstage/core-compat-api      | package.json:58:6 | error    |
-| @backstage/errors               | package.json:61:6 | error    |
-| victory                         | package.json:89:6 | error    |
+| @backstage/plugin-user-settings | package.json:76:6 | error    |
+| @backstage/filter-predicates    | package.json:66:6 | error    |
+| @backstage/core-compat-api      | package.json:62:6 | error    |
+| @backstage/errors               | package.json:65:6 | error    |
+| victory                         | package.json:93:6 | error    |
