+++ workspaces/topology/plugins/topology/package.json
@@ -17,17 +17,21 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./alpha": "./src/alpha/index.tsx",
-    "./translations": "./src/translations/exports.ts",
+    "./alpha": "./src/alpha.ts",
+    "./legacy": "./src/legacy.ts",
+    "./translations": "./src/translations/index.ts",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
-        "src/alpha/index.tsx"
+        "src/alpha.ts"
+      ],
+      "legacy": [
+        "src/legacy.ts"
       ],
       "translations": [
-        "src/translations/exports.ts"
+        "src/translations/index.ts"
       ],
       "package.json": [
         "package.json"
@@ -47,8 +51,8 @@
     "prepack": "backstage-cli package prepack",
     "start": "backstage-cli package start",
     "start:mock": "backstage-cli package start --entrypoint dev/index.mock.tsx",
-    "start:alpha": "backstage-cli package start --entrypoint dev/alpha/index.tsx",
-    "start:alpha:mock": "backstage-cli package start --entrypoint dev/alpha/index.mock.tsx",
+    "start:legacy": "backstage-cli package start --entrypoint dev/legacy.tsx",
+    "start:legacy:mock": "backstage-cli package start --entrypoint dev/legacy.mock.tsx",
     "test": "backstage-cli package test --passWithNoTests --coverage",
     "tsc": "tsc"
   },
@@ -120,6 +124,14 @@
     "dist",
     "app-config.backstage-community.yaml"
   ],
+  "scalprum": {
+    "name": "backstage-community.plugin-topology",
+    "exposedModules": {
+      "PluginRoot": "./src/index.ts",
+      "Legacy": "./src/legacy.ts",
+      "Alpha": "./src/alpha.ts"
+    }
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/backstage/community-plugins",
