+++ workspaces/topology/plugins/topology/report-alpha.api.md
@@ -3,17 +3,9 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { Entity } from '@backstage/catalog-model';
-import { FrontendPlugin } from '@backstage/frontend-plugin-api';
+import { FrontendModule } from '@backstage/frontend-plugin-api';
 import { TranslationRef } from '@backstage/frontend-plugin-api';
 import { TranslationResource } from '@backstage/frontend-plugin-api';
-
-// @alpha
-export const isTopologyAvailable: (entity: Entity) => boolean;
-
-// @alpha
-const topologyPlugin: FrontendPlugin;
-export default topologyPlugin;
 
 // @alpha
 export const topologyTranslationRef: TranslationRef<
@@ -156,5 +148,6 @@ export const topologyTranslationRef: TranslationRef<
 // @alpha
 export const topologyTranslations: TranslationResource<'plugin.topology'>;
 
-// (No @packageDocumentation comment for this package)
+// @public
+export const topologyTranslationsModule: FrontendModule;
 ```
