+++ workspaces/topology/plugins/topology/report-legacy.api.md
@@ -3,13 +3,18 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { BackstagePlugin } from '@backstage/core-plugin-api';
 import { Entity } from '@backstage/catalog-model';
-import { FrontendPlugin } from '@backstage/frontend-plugin-api';
+import { JSX as JSX_2 } from 'react/jsx-runtime';
 
 // @public
 export const isTopologyAvailable: (entity: Entity) => boolean;
 
 // @public
-const topologyPlugin: FrontendPlugin;
-export default topologyPlugin;
+export const TopologyPage: () => JSX_2.Element;
+
+// @public
+export const topologyPlugin: BackstagePlugin<{}, {}, {}>;
+
+// (No @packageDocumentation comment for this package)
 ```
