+++ workspaces/topology/plugins/topology/src/alpha.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,10 +15,8 @@
  */
 
 /**
- * Topology plugin for Backstage.
- *
+ * @deprecated Use the root import instead: `import from '@backstage-community/plugin-topology'`
  * @packageDocumentation
  */
 
-export { default } from './plugin';
-export { isTopologyAvailable } from './isTopologyAvailable';
+export * from './translations';
