+++ workspaces/topology/plugins/topology/src/legacy.test.ts
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { topologyPlugin } from './legacy';
 
-/**
- * Topology plugin for Backstage.
- *
- * @packageDocumentation
- */
-
-export { default } from './plugin';
-export { isTopologyAvailable } from './isTopologyAvailable';
+describe('topology legacy', () => {
+  it('should export plugin', () => {
+    expect(topologyPlugin).toBeDefined();
+    expect(topologyPlugin.getId()).toBe('topology');
+  });
+});
