+++ workspaces/topology/plugins/topology/src/plugin.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { topologyPlugin } from './plugin';
+import { Entity } from '@backstage/catalog-model';
+
+import topologyPlugin from './plugin';
+import { isTopologyAvailable } from './isTopologyAvailable';
 
 describe('topology', () => {
-  it('should export plugin', () => {
+  it('should export the new frontend system plugin', () => {
     expect(topologyPlugin).toBeDefined();
+    expect(topologyPlugin.pluginId).toBe('topology');
+  });
+
+  it('isTopologyAvailable is true when kubernetes id or namespace annotations are set', () => {
+    expect(
+      isTopologyAvailable({
+        apiVersion: 'backstage.io/v1alpha1',
+        kind: 'Component',
+        metadata: {
+          name: 'example',
+          annotations: { 'backstage.io/kubernetes-id': 'example' },
+        },
+      } as Entity),
+    ).toBe(true);
+    expect(
+      isTopologyAvailable({
+        apiVersion: 'backstage.io/v1alpha1',
+        kind: 'Component',
+        metadata: {
+          name: 'example',
+          annotations: { 'backstage.io/kubernetes-namespace': 'default' },
+        },
+      } as Entity),
+    ).toBe(true);
+    expect(
+      isTopologyAvailable({
+        apiVersion: 'backstage.io/v1alpha1',
+        kind: 'Component',
+        metadata: { name: 'example' },
+      } as Entity),
+    ).toBe(false);
   });
 });
