--- workspaces/topology/plugins/topology/src/translations/exports.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright 2026 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createFrontendModule } from '@backstage/frontend-plugin-api';
-import { TranslationBlueprint } from '@backstage/plugin-app-react';
-
-import { topologyTranslations } from './index';
-
-export { topologyTranslations } from './index';
-export { topologyTranslationRef } from './ref';
-
-/**
- * App module that automatically registers the topology plugin translations.
- *
- * @alpha
- */
-const topologyTranslationsModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [
-    TranslationBlueprint.make({
-      name: 'topology-translations',
-      params: {
-        resource: topologyTranslations,
-      },
-    }),
-  ],
-});
-
-export default topologyTranslationsModule;
